Split ProgressBarTimer from ProgressBar
Split the `<ProgressBar>` component into two separate components:
`<ProgressBar>` and `<ProgressBarTimer>`.
