+++ docs/ProgressBar.mdx
@@ -9,9 +9,7 @@ import { ProgressBar } from '../src'
 
 <Playground>
   <ProgressBar
-    countdownStart="2019-01-01T12:00:00.000-04:00"
-    countdownEnd="2019-01-14T12:00:00.000-04:00"
-    currentTime="2019-01-05T12:00:00.000-04:00"
+    percentComplete={70}
     highlight="purple100"
   />
 </Playground>
