+++ docs/ProgressBarTimer.mdx
@@ -1,0 +1,17 @@
+---
+name: ProgressBarTimer
+---
+
+import { Playground } from 'docz'
+import { ProgressBarTimer } from '../src'
+
+# ProgressBarTimer
+
+<Playground>
+  <ProgressBarTimer
+    countdownStart="2019-01-01T12:00:00.000-04:00"
+    countdownEnd="2019-01-14T12:00:00.000-04:00"
+    currentTime="2019-01-05T12:00:00.000-04:00"
+    highlight="purple30"
+  />
+</Playground>
