+++ src/elements/ProgressBar/ProgressBar.tsx
@@ -1,4 +1,3 @@
-import moment from "moment"
 import React from "react"
 import styled from "styled-components"
 
@@ -13,33 +12,16 @@ const ProgressBarBackground = styled.div`
 `
 
 export const ProgressBar: React.SFC<{
-  currentTime: string
-  countdownStart: string
-  countdownEnd: string
+  percentComplete: number
   highlight: Parameters<typeof color>[0]
-}> = ({
-  currentTime,
-  countdownStart,
-  countdownEnd,
-  highlight = "purple100",
-}) => {
-  const secondsRemaining = moment(countdownEnd).diff(
-    moment(currentTime),
-    "seconds"
-  )
-  const totalSeconds = moment(countdownEnd).diff(
-    moment(countdownStart),
-    "seconds"
-  )
-  const progress = Math.max(0, (secondsRemaining * 100) / totalSeconds)
-
+}> = ({ percentComplete, highlight = "purple100" }) => {
   return (
     <ProgressBarBackground>
       <div
         style={{
           transition: "width 0.34s ease",
           backgroundColor: color(highlight as any),
-          width: progress + "%",
+          width: Math.round(percentComplete) + "%",
         }}
       />
     </ProgressBarBackground>
