+++ src/elements/index.tsx
@@ -16,6 +16,7 @@ export * from "./Pagination"
 export * from "./Link"
 export * from "./PriceRange"
 export * from "./ProgressBar"
+export * from "./ProgressBarTimer"
 export * from "./Radio"
 export * from "./RadioGroup"
 export * from "./Select"
